pyproject.toml
[tool.poetry]
name = "smac-fastapi-auth"
version = "0.1.0"
description = ""
authors = ["Sven Maschek <dev@smaschek.com>"]
readme = "README.md"
packages = [{include = "smac_fastapi_auth", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.3.4"
fastapi-cognito = "^2.5.0"
loguru = "^0.7.2"
aiohttp = "^3.9.5"
cachebox = "^3.2.1"
boto3 = "^1.34.139"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
pytest = "^8.2.2"
pre-commit = "^3.7.1"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"
gevent = "^24.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
extend-exclude = ["~/miniforge3"]
extend-include = ["*.ipynb"]

[tool.ruff.lint.isort]
order-by-type = true
relative-imports-order = "closest-to-furthest"
extra-standard-library = ["typing"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
known-first-party = []

[tool.pytest.ini_options]
addopts = "--cov='smac_fastapi_auth' --junitxml=junit/test-results.xml --cov=com --cov-report=xml --cov-report html --cov-report term-missing --cov-fail-under 50"

[tool.coverage.run]
parallel = true
branch = true
concurrency = ["gevent"]
omit = [
    "/tests/*",
    "*/tests/*",
    "*/test/*",
    "*/__init__.py",
    "*/__main__.py",
    "*/setup.py",
    "*/conftest.py",
    "*/config.py",
    "*/migrations/*",
    "*/migrations.py",
    "*/*_main.py",
    "constants.py",
]

[tool.coverage.report]
exclude_also = [
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
